bits of refactoring
