+++ src/plonk/better_cs/mod.rs
@@ -3,5 +3,6 @@ pub mod test_assembly;
 pub mod adaptor;
 pub mod gates;
 pub mod data_structures;
+pub mod utils;
 pub mod generators;
-//pub mod prover;
+pub mod prover;
